Change location for drupal.inc template.
There is the problem if drupal.inc configuration hosted in /etc/nginx/conf.d. Those configs are imported every time and this cause problem when not existed sited on subdomain is open first site found, except default domain with boilerplate. This is also can provide problem when opens wrong site.